refactor tests for pystiche.image (#378)
* refactor tests for pystiche.image

* refactor tests for pystiche.image.transforms

* delete old

* cleanup
