+++ tox.ini
@@ -13,11 +13,12 @@ passenv =
   GITHUB_ACTIONS
 deps =
   pytest >= 6
-  pyimagetest < 0.3
+  pyimagetest >= 0.3
   pillow_affine
   dill
   pytest-subtests >= 0.3.2
   pytest-cov
+  pytest-mock >= 3.1
   # TODO: move to a released version
   git+https://github.com/pmeier/pytorch_testing_utils
 commands =
